pyproject.toml
[tool.poetry]
name = "smart-spend"
version = "0.1.0"
description = "Smart Spend - приложение для управления расходами"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.scripts]
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000"

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = "0.104.1"
uvicorn = { extras = ["standard"], version = "0.24.0" }
pydantic = { extras = ["email"], version = "2.5.0" }
pydantic-settings = "2.1.0"
sqlalchemy = "2.0.23"
alembic = "1.12.1"
asyncpg = "0.29.0"
python-dotenv = "1.0.0"
python-multipart = "0.0.6"
python-jose = { extras = ["cryptography"], version = "3.3.0" }
passlib = { extras = ["bcrypt", "argon2"], version = "1.7.4" }
argon2-cffi = "23.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
pytest-asyncio = "0.21.1"
httpx = "0.25.2"
black = "23.11.0"
ruff = "0.1.6"
mypy = "1.7.1"
pre-commit = "3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true
# no_implicit_optional = true  # По умолчанию включено в новых версиях mypy
# strict_optional = true  # По умолчанию включено
